freqhopping: Fix clang `void-pointer-to-int-cast` warning
../drivers/misc/mediatek/freqhopping/fhctl_new/clk-fhctl-mcupm.c:308:6:
 error: cast to smaller integer type 'unsigned int' from 'void *'
 [-Werror,-Wvoid-pointer-to-int-cast]
 + (unsigned int)match_data->reg_tr;
 ^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~

+++ drivers/misc/mediatek/freqhopping/fhctl_new/clk-fhctl-mcupm.c
@@ -309,7 +309,7 @@ static int mcupm_init_v1(struct pll_dts *array, struct match *match)
 	match_data = match->hdlr->data;
 	if (match_data && match_data->reg_tr) {
 		priv_data->reg_tr = array->fhctl_base
-			+ (unsigned int)match_data->reg_tr;
+			+ (uintptr_t)match_data->reg_tr;
 	}
 
 	/* hook to array */
